fix: incorrect apple-touch-icon links
We were getting some 404s given that we were referencing a file that
does not exist. This commit changes it so that we actually add a valid
`apple-touch-icon.png` and `apple-touch-icon-precomposed.png` so that
no iOS device will cause a 404s when trying to request the icon.

Worth noting that the icon looks good now even before this change and
– it's probably fetching the icon from the favicon or something – and
removing the link does not change that but it will generate a lot of
404s.

Though the dimensions of the icon can vary between iOS devices,
including a link tag with a `sizes` of `180x180` in the `<head>` of the
page is enough.

References:

* https://github.com/h5bp/html5-boilerplate/blob/29d7d80/docs/extend.md#apple-touch-icons
* https://developer.apple.com/library/archive/documentation/AppleApplications/Reference/SafariWebContent/ConfiguringWebApplications/ConfiguringWebApplications.html
* https://mathiasbynens.be/notes/touch-icons