infra: use `--primary` flag to run the load YAML commands
This flag will ensure that this command is only run in the `primary`
host, meaning that it won't be triggered multiple times like it would be
if we don't add this flag.

We don't actually have multiple hosts, but this is just a safety measure
in case we add multiple hosts in the future and we forgot to update this
task.