ci: prevent brakeman from failing when it's outdated
The CI is failing because `brakeman` is outdated:

```
Run bin/brakeman --no-pager
Brakeman 6.1.2 is not the latest version 6.2.1
```

Let's avoid blocking people from merging simply because brakeman is outdated. We update our dependencies every week so we tend to be up-to-date, therefore even if we don't have the latest version we still have a maximum 1-week-old release which should be enough to catch most of the potential vulnerabilities. 

If not, we will catch it pretty quickly as brakeman will be updated in short time and the CI will fail if it detectes a vulnerability.