cd: increase Kamal's `max_attempts`
We were facing an issue when deploying where the deploys, after starting
the new container with the new version of the app, failed to remove the
old container as Kamal reached the `max_attemps` before the container
was marked as `unhealthy` by Docker.

After digging around for a bit, we realized that the issue is that
when we increased the `interval` of the healthchecks, we didn't Docker
enough time to mark the container as unhealthy before Kamal reached the
`max_attempts` inumber of attempt and thus failed the deploy.

For context, Docker marks the container as `unhealthy` after 3 consecutive
failed health checks. In our case, where we run one Docker healthcheck
every 30s, that means that it takes at least 90s for Docker to mark the
container as unhealthy.

On the other hand, when deploying, Kamal checks if the service is
healthy for an at-most `max_attempts` number of times, where each check
is separated by an increasing interval starting with 1 second. In our
case, we set `max_attempts` to `8` so it will take Kamal 1+2+3+...+7=28
to reach it's limit of attempts.

Therefore, the issue is that Kamal reaches its limit of attempts before
Docker gets to even try a third healthcheck, thus not marking the old
container as unhealthy and making the deploy to fail.

This commit fixes this issue by increasing the `max_attempts` to 14
which means that it will take Kamal `1+2+...+12+13=91s` to reach its
limit which should be enough time for three unsuccessful Docker
healthchecks occur.