fix(test): ensure that sign in request succeeded before continuing
The test was failing with the following error:

```
Error:
UserTest#test_can_edit_profile:
Capybara::ElementNotFound: Unable to find visible link or button "Editar Perfil"
    test/system/user_test.rb:135:in 'block in <class:UserTest>'
```

The issue is that there's a race condition between the application
loading the home page and capybara clicking on the user menu and it
having the `Editar Perfil` link. If capybara is too fast, it will click
on the user menu before the application loads so the click won't open
the menu and thus it won't find the button.

In order to fix this, we can wait until the text `Iniciaste sesión
correctamente` appears in the page, which will indicate when we can
continue with the rest of the actions.