test: fix system specs flakiness
It seems that, even though it's running in headless mode, chrome is
somehow deciding that its window is "covered up" by other windows and
it's treating the foreground window as backgrounded, which causes it to
halt animations, resulting in tests errors.

In order to fix this, we can pass the
`disable-backgrounding-occluded-windows` flag to the driver so that
windows covered by other windows are not treated as backgrounded.

Hopefully this is only a temporary fix while we wait for
`chromedriver`/`chromium` to fix the issue.

References:
- https://github.com/teamcapybara/capybara/issues/2800
- https://github.com/teamcapybara/capybara/issues/2796