feat: prevent banner for hiding content
The previous banner from MaterialUI that we were using was able to
render without hiding the rest of the content of the page.

This commits makes the new banner behave in a similar way, following a
similar approach than the one using before: using a relative div with
the same heigh than the banner for reserving the space that the banner
would take. This is in turn achieved by using Javascript to calculate
the height of the banner and setting the height of the relative div to
be the same.