feat: migrate welcome banner to Tailwind (#775)
* feature: migrate banner to tailwind

* feat: prevent banner for hiding content

The previous banner from MaterialUI that we were using was able to
render without hiding the rest of the content of the page.

This commits makes the new banner behave in a similar way, following a
similar approach than the one using before: using a relative div with
the same heigh than the banner for reserving the space that the banner
would take. This is in turn achieved by using Javascript to calculate
the height of the banner and setting the height of the relative div to
be the same.

* chore: unpin `@material/banner`

Run `bin/importmap unpin @material/banner`

* chore: add needed importmaps pins again

The user menu stills depends on them

* chore: remove unnecessary classes

* chore: change target name to make the code clearer

* feat: use `text-pretty` instead of `text-balance`

`text-pretty` looks better in phones with big screens as the iPhone Pro
Max.

* chore: no need to specify that we are listening for the `submit` event

This is the default event in forms.

* feat: allow navbar to withstand window being resized

* build: remove unnecessary `importmap` pins

We weren't able to remove before as the user menu still depended on
them, but now that the user menu was migrated out of MaterialUI in #745,
we can remove these as part of this PR since the banner is the last
component that uses them.

* chore: remove now-unused css classes

* feature: adjust `top` css property to match new navbar

* feat: constraint banner content to the same width as the main content

* feat: use uppercase letters for banner button