feat(planner): use drag and drop to move subjects from one semester to another (#842)
* chore: run `bin/importmap pin sortablejs`

* feat: add `draggable` controller

* feat: save the change of semester to the database

* chore: move no subjects planned message outside div

This was causing problems were sometimes the text would appear above the
'add subject to semester section' when adding an element to an empty
semester.

* fix: semester appearing over the footer

* feat: disable sorting elements on the same semester

* feat: allow dragging not approved subjects

* refactor: import `FetchRequest`

* feat: remove select for adding not planned subjects to semester

* feat: make not planned approved subjects have same spacing as planned subjects

This was causing trouble as dragging and dropping a not planned approved
subject into a semester will look weird until the page is reloaded.

* feat(planner): use turbo to reload section

This will allow us to refresh the credits from each semester when moving
subjects from one to another. What's more, it also allow us to add the
`remove` button to a subject when moving it from the not-planned
approved subjects section into another semester.

* feat: use an icon for moving the elements

* test: fix specs to work with drag and drop

* fix: subject selectors breaking when adding/removing a subject

* fix: semester param not being sent correctly

* fix: semesters' credits not updating

* chore: use strong params

* fix: draggable controller having the wrong scope

The div incorrect closing `div` was closing the controller's div before
enclosing the planner semesters

* chore: use stimulus targets

* chore: improve readability

* test: fix broken spec

* chore: group semester card sections inside semester card folder

* chore: fix wrong instance var name

* refactor: rename `draggable` controller to `planner-draggable`

* feat: reload page if something fails

* test: moving a subject from one semester to another

* refactor: avoid adding unnecessary `li`

* chore: fix incorrect target name

* style: use hash shorthand syntax

* refactor: avoid wrapping subject rows with form for destroy action

This reverts commit cbaae36d15766aecd34d01884368243dd6ff5019 and moves
the form for the remove button to just wrap said button by using
`button_to`.

---------

Co-authored-by: Nicolas Temciuc <nicolas.temciuc@cedarcode.com>