feat: only search by `label`
The default in `choices.js` is to search by `["label", "value"]`.

This was causing some strange behaviour since the values are the ids
from the subject which are different from the codes.

This change will also help to search by subject code since the code is
in the label.

See:
https://github.com/Choices-js/Choices?tab=readme-ov-file#searchfields