fix: absolute div making body to not take full width of the page
The absolute `div` used for centering the credits text when outside of
the progress bar was taking more width than its parent, making the whole
page take more right space compared to the viewport so it was not
looking good on mobile. Adding the `relative` position to the parent
makes its `absolute` child only take the width of its parent.